requirements.txt
color_matcher==0.5.0
colour==0.1.5
numpy==2.2.3
Pillow==9.4.0
Pillow==11.1.0
pillow_lut==1.0.1
